seller/requirements.txt
Flask==3.0.0
mysql-connector-python==9.0.0
boto3==1.35.0
Werkzeug==3.0.1
gunicorn==20.1.0